Move Lambda inst into it's own package
