+++ instana/__init__.py
@@ -108,7 +108,7 @@ def boot_agent():
     # Instrumentation
     if "INSTANA_DISABLE_AUTO_INSTR" not in os.environ:
         # Import & initialize instrumentation
-        from .instrumentation import aws_lambda
+        from .instrumentation.aws import lambda_inst
 
         if sys.version_info >= (3, 5, 3):
             from .instrumentation import asyncio
